Commit #5 CSE 27MAY2022 - event handlers and textfield added to timers. condition: when a timer button is clicked. execution: A variable, timer gets assigned the value of the button.
-event handler added to power button
condition -when both the target temp is set and a timer button has been
clicked (timer variable has a value that's not 0 assigned)
execution: actual temp is set equal to target temp and a timer starts.
when the timer is done, the actual temp reverts back to 68F

Warning: power button will throw an exception if clicked without conditions being met!